refactor: rename common to shared
