+++ src/lib.rs
@@ -2,7 +2,7 @@
 
 extern crate alloc;
 
-pub mod common;
+pub mod shared;
 pub mod sync;
 pub mod unsync;
 // pub mod list;
